+++ src/components/AllFolderLinks.css
@@ -4,6 +4,7 @@
   flex-wrap: wrap;
   justify-content: space-between;
   padding: 0;
+  width: 100%;
 }
 
 .folder-card .link-card {
@@ -87,11 +88,6 @@
 }
 
 @media only screen and (max-width: 1124px) {
-  .folder-card {
-    width: 800px;
-    margin: 0 auto;
-  }
-
   .folder-card .link-card {
     width: calc(50% - 20px);
   }
